gameboard and cell components unified
